services implemented

+++ app/services/ExpenseService.scala
@@ -8,8 +8,8 @@ import scala.concurrent.Future
 @ImplementedBy(classOf[ExpenseServiceImpl])
 trait ExpenseService {
   def addExpense(expense: Expense): Future[String]
-  def deleteExpense(id: Long): Future[Int]
-  def getExpense(id: Long): Future[Option[Expense]]
+  def deleteExpense(id: Int): Future[Int]
+  def getExpense(id: Int): Future[Option[Expense]]
   def listAllExpenses: Future[Seq[Expense]]
-  def updateExpense(id: Long, expense: Expense) : Future[String]
+  def updateExpense(id: Int, expense: Expense) : Future[String]
 }
